+++ services/workflows-service/src/app.module.ts
@@ -45,6 +45,8 @@ import { ScheduleModule } from '@nestjs/schedule';
 import { CronModule } from '@/workflow/cron/cron.module';
 import z from 'zod';
 import { hashKey } from './customer/api-key/utils';
+import { RuleEngineModule } from './rule-engine/rule-engine.module';
+import { NotionModule } from '@/notion/notion.module';
 
 export const validate = async (config: Record<string, unknown>) => {
   const zodEnvSchema = z
@@ -120,6 +122,8 @@ export const validate = async (config: Record<string, unknown>) => {
     CronModule,
     ScheduleModule.forRoot(),
     initHttpMoudle(),
+    RuleEngineModule,
+    NotionModule,
   ],
   providers: [
     {