+++ services/workflows-service/src/collection-flow/controllers/collection-flow.end-user.controller.ts
@@ -1,7 +1,7 @@
 import { CollectionFlowService } from '@/collection-flow/collection-flow.service';
 import { TokenScope, type ITokenScope } from '@/common/decorators/token-scope.decorator';
 import { UseTokenAuthGuard } from '@/common/guards/token-guard/use-token-auth.decorator';
-import { EndUserCreateDto } from '@/end-user/dtos/end-user-create';
+import { EndUserUpdateDto } from '@/end-user/dtos/end-user-update';
 import { EndUserModel } from '@/end-user/end-user.model';
 import { EndUserService } from '@/end-user/end-user.service';
 import * as common from '@nestjs/common';
@@ -19,7 +19,7 @@ export class CollectionFlowEndUserController {
 
   @common.Post()
   @swagger.ApiCreatedResponse({ type: [EndUserModel] })
-  getCompanyInfo(@TokenScope() tokenScope: ITokenScope, @common.Body() data: EndUserCreateDto) {
+  getCompanyInfo(@TokenScope() tokenScope: ITokenScope, @common.Body() data: EndUserUpdateDto) {
     return this.endUserService.updateById(tokenScope.endUserId, { data: data });
   }
 }