+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/ui
 
+## 0.5.6
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.12
+
 ## 0.5.5
 
 ### Patch Changes