+++ services/websocket-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/websocket-service
 
+## 0.1.5
+
+### Patch Changes
+
+- Bump
+
 ## 0.1.4
 
 ### Patch Changes